+++ CotEditor/Sources/ToolbarController.swift
@@ -34,46 +34,32 @@ final class ToolbarController: NSObject {
     weak var document: Document? {
         
         willSet {
-            self.documentObservers.removeAll()
+            self.documentStyleObserver = nil
         }
         
         didSet {
             guard let document = document else { return }
             
-            self.invalidateLineEndingSelection(to: document.lineEnding)
-            self.invalidateEncodingSelection()
-            self.invalidateSyntaxStyleSelection()
             self.toolbar?.validateVisibleItems()
+            self.invalidateSyntaxStyleSelection()
             
-            // observe document status change
-            document.didChangeEncoding
-                .receive(on: RunLoop.main)
-                .sink { [weak self] _ in self?.invalidateEncodingSelection() }
-                .store(in: &self.documentObservers)
-            document.didChangeLineEnding
-                .receive(on: RunLoop.main)
-                .sink { [weak self] in self?.invalidateLineEndingSelection(to: $0) }
-                .store(in: &self.documentObservers)
-            document.didChangeSyntaxStyle
+            // observe document's style change
+            self.documentStyleObserver = document.didChangeSyntaxStyle
                 .receive(on: RunLoop.main)
                 .sink { [weak self] _ in self?.invalidateSyntaxStyleSelection() }
-                .store(in: &self.documentObservers)
         }
     }
     
     
     // MARK: Private Properties
     
-    private var documentObservers: Set<AnyCancellable> = []
-    private var menuUpdateObservers: Set<AnyCancellable> = []
+    private var documentStyleObserver: AnyCancellable?
+    private var styleListObserver: AnyCancellable?
     private var recentStyleNamesObserver: UserDefaultsObservation?
     
     @IBOutlet private weak var toolbar: NSToolbar?
-    @IBOutlet private weak var lineEndingPopupButton: NSPopUpButton?
-    @IBOutlet private weak var encodingPopupButton: NSPopUpButton?
-    @IBOutlet private weak var syntaxPopupButton: NSPopUpButton?
-    
     @IBOutlet private weak var shareToolbarItem: NSToolbarItem?
+    @IBOutlet private weak var syntaxPopupButton: NSPopUpButton?
     
     
     
@@ -84,23 +70,17 @@ final class ToolbarController: NSObject {
         
         super.awakeFromNib()
         
-        self.buildEncodingPopupButton()
-        self.buildSyntaxPopupButton()
-        
         // setup Share toolbar item
         // -> The Share button action must be called on `mouseDown`.
         (self.shareToolbarItem!.view as! NSButton).sendAction(on: .leftMouseDown)
         self.shareToolbarItem!.menuFormRepresentation = NSDocumentController.shared.standardShareMenuItem()
         
-        // observe popup menu line-up change
-        self.menuUpdateObservers.removeAll()
-        EncodingManager.shared.didUpdateSettingList
-            .sink { [weak self] _ in self?.buildEncodingPopupButton() }
-            .store(in: &self.menuUpdateObservers)
-        SyntaxManager.shared.didUpdateSettingList
-            .sink { [weak self] _ in self?.buildSyntaxPopupButton() }
-            .store(in: &self.menuUpdateObservers)
+        // setup syntax style menu
+        self.buildSyntaxPopupButton()
         
+        //  observe for syntax style line-up change
+        self.styleListObserver = SyntaxManager.shared.didUpdateSettingList
+            .sink { [weak self] _ in self?.buildSyntaxPopupButton() }
         self.recentStyleNamesObserver?.invalidate()
         self.recentStyleNamesObserver = UserDefaults.standard.observe(key: .recentStyleNames) { [weak self] _ in
             self?.buildSyntaxPopupButton()
@@ -110,27 +90,6 @@ final class ToolbarController: NSObject {
     
     
     // MARK: Private Methods
-    
-    /// select item in the encoding popup menu
-    private func invalidateLineEndingSelection(to lineEnding: LineEnding) {
-        
-        self.lineEndingPopupButton?.selectItem(withTag: lineEnding.index)
-    }
-    
-    
-    /// select item in the line ending menu
-    private func invalidateEncodingSelection() {
-        
-        guard let fileEncoding = self.document?.fileEncoding else { return }
-        
-        var tag = Int(fileEncoding.encoding.rawValue)
-        if fileEncoding.withUTF8BOM {
-            tag *= -1
-        }
-        
-        self.encodingPopupButton?.selectItem(withTag: tag)
-    }
-    
     
     /// select item in the syntax style menu
     private func invalidateSyntaxStyleSelection() {
@@ -142,17 +101,6 @@ final class ToolbarController: NSObject {
         if popUpButton.selectedItem == nil {
             popUpButton.selectItem(at: 0)  // select "None"
         }
-    }
-    
-    
-    /// build encoding popup item
-    private func buildEncodingPopupButton() {
-        
-        guard let popUpButton = self.encodingPopupButton else { return }
-        
-        EncodingManager.shared.updateChangeEncodingMenu(popUpButton.menu!)
-        
-        self.invalidateEncodingSelection()
     }